Make byte-code an optional feature
