+++ crates/rune-macros/Cargo.toml
@@ -15,7 +15,7 @@ categories = ["parser-implementations"]
 
 [dependencies]
 rune-core = { version = "=0.12.3", path = "../rune-core" }
-syn = { version = "2.0.15", features = ["full"] }
+syn = { version = "2.0.16", features = ["full"] }
 quote = "1.0.27"
 proc-macro2 = { version = "1.0.56", features = ["span-locations"] }
