+++ crates/rune/src/runtime/unit/storage.rs
@@ -5,6 +5,7 @@ use core::slice;
 use crate::no_std::error;
 use crate::no_std::vec::Vec;
 
+#[cfg(feature = "byte-code")]
 use musli_storage::error::BufferError;
 use serde::{Deserialize, Serialize};
 
@@ -13,38 +14,45 @@ use crate::runtime::Inst;
 mod sealed {
     pub trait Sealed {}
 
-    impl Sealed for super::BytesStorage {}
-
-    impl Sealed for super::ArrayStorage {}
+    #[cfg(feature = "byte-code")]
+    impl Sealed for crate::runtime::unit::ByteCodeUnit {}
+    impl Sealed for crate::runtime::unit::ArrayUnit {}
 }
 
 /// Builder trait for unit storage.
-pub trait UnitStorageBuilder: self::sealed::Sealed {
+pub trait UnitEncoder: self::sealed::Sealed {
     /// Current offset in storage, which also corresponds to the instruction
     /// pointer being built.
+    #[doc(hidden)]
     fn offset(&self) -> usize;
 
     /// Encode an instruction into the current storage.
+    #[doc(hidden)]
     fn encode(&mut self, inst: Inst) -> Result<(), EncodeError>;
 
     /// Indicate that the given number of offsets have been added.
+    #[doc(hidden)]
     fn extend_offsets(&mut self, extra: usize) -> usize;
 
     /// Mark that the given offset index is at the current offset.
+    #[doc(hidden)]
     fn mark_offset(&mut self, index: usize);
 
     /// Calculate label jump.
+    #[doc(hidden)]
     fn label_jump(&self, base: usize, offset: usize, jump: usize) -> usize;
 }
 
 /// Instruction storage used by a [`Unit`][super::Unit].
-pub trait UnitStorage:
-    self::sealed::Sealed + UnitStorageBuilder + fmt::Debug + Clone + Default
-{
+pub trait UnitStorage: self::sealed::Sealed + fmt::Debug + Default + Clone {
     /// Iterator over instructions and their corresponding instruction offsets.
     type Iter<'this>: Iterator<Item = (usize, Inst)>
     where
         Self: 'this;
+
+    /// Size of unit storage. This can be seen as the instruction pointer which
+    /// is just beyond the last instruction.
+    fn len(&self) -> usize;
 
     /// Iterate over all instructions.
     fn iter(&self) -> Self::Iter<'_>;
@@ -56,107 +64,13 @@ pub trait UnitStorage:
     fn translate(&self, jump: usize) -> Result<usize, BadJump>;
 }
 
-/// Unit stored as bytes.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct BytesStorage {
-    /// The instructions contained in the source file.
-    bytes: Vec<u8>,
-    /// Known jump offsets.
-    offsets: Vec<usize>,
-}
-
-/// Iterator for [`BytesStorage`].
-pub struct BytesStorageIter<'a> {
-    address: &'a [u8],
-    len: usize,
-}
-
-impl<'a> Iterator for BytesStorageIter<'a> {
-    type Item = (usize, Inst);
-
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.address.is_empty() {
-            return None;
-        }
-
-        let ip = self.len.checked_sub(self.address.len())?;
-        let inst = musli_storage::decode(&mut self.address).ok()?;
-        Some((ip, inst))
-    }
-}
-
-impl UnitStorageBuilder for BytesStorage {
-    #[inline]
-    fn offset(&self) -> usize {
-        self.bytes.len()
-    }
-
-    #[inline]
-    fn encode(&mut self, inst: Inst) -> Result<(), EncodeError> {
-        musli_storage::encode(&mut self.bytes, &inst)?;
-        Ok(())
-    }
-
-    #[inline]
-    fn extend_offsets(&mut self, extra: usize) -> usize {
-        let base = self.offsets.len();
-        self.offsets.extend((0..extra).map(|_| 0));
-        base
-    }
-
-    #[inline]
-    fn mark_offset(&mut self, index: usize) {
-        if let Some(o) = self.offsets.get_mut(index) {
-            *o = self.bytes.len();
-        }
-    }
-
-    #[inline]
-    fn label_jump(&self, base: usize, _: usize, jump: usize) -> usize {
-        base.wrapping_add(jump)
-    }
-}
-
-impl UnitStorage for BytesStorage {
-    type Iter<'this> = BytesStorageIter<'this>;
-
-    #[inline]
-    fn iter(&self) -> Self::Iter<'_> {
-        BytesStorageIter {
-            address: &self.bytes[..],
-            len: self.bytes.len(),
-        }
-    }
-
-    fn get(&self, ip: usize) -> Result<Option<(Inst, usize)>, BadInstruction> {
-        let Some(mut bytes) = self.bytes.get(ip..) else {
-            return Ok(None);
-        };
-
-        let start = bytes.as_ptr();
-        let inst: Inst = musli_storage::decode(&mut bytes).map_err(|_| BadInstruction { ip })?;
-        let len = (bytes.as_ptr() as usize).wrapping_sub(start as usize);
-        Ok(Some((inst, len)))
-    }
-
-    #[inline]
-    fn translate(&self, jump: usize) -> Result<usize, BadJump> {
-        let Some(&offset) = self.offsets.get(jump) else {
-            return Err(BadJump { jump });
-        };
-
-        Ok(offset)
-    }
-}
-
 /// Unit stored as array of instructions.
 #[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct ArrayStorage {
+pub struct ArrayUnit {
     instructions: Vec<Inst>,
 }
 
-impl UnitStorageBuilder for ArrayStorage {
+impl UnitEncoder for ArrayUnit {
     #[inline]
     fn offset(&self) -> usize {
         self.instructions.len()
@@ -182,8 +96,13 @@ impl UnitStorageBuilder for ArrayStorage {
     }
 }
 
-impl UnitStorage for ArrayStorage {
+impl UnitStorage for ArrayUnit {
     type Iter<'this> = iter::Enumerate<iter::Copied<slice::Iter<'this, Inst>>>;
+
+    #[inline]
+    fn len(&self) -> usize {
+        self.instructions.len()
+    }
 
     #[inline]
     fn iter(&self) -> Self::Iter<'_> {
@@ -207,10 +126,12 @@ impl UnitStorage for ArrayStorage {
 
 /// Error indicating that encoding failed.
 #[derive(Debug)]
+#[doc(hidden)]
 pub struct EncodeError {
     kind: EncodeErrorKind,
 }
 
+#[cfg(feature = "byte-code")]
 impl From<BufferError> for EncodeError {
     #[inline]
     fn from(error: BufferError) -> Self {
@@ -224,6 +145,7 @@ impl fmt::Display for EncodeError {
     #[inline]
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match &self.kind {
+            #[cfg(feature = "byte-code")]
             EncodeErrorKind::BufferError { error } => error.fmt(f),
         }
     }
@@ -233,6 +155,7 @@ impl error::Error for EncodeError {}
 
 #[derive(Debug)]
 enum EncodeErrorKind {
+    #[cfg(feature = "byte-code")]
     BufferError { error: BufferError },
 }
 
@@ -240,7 +163,7 @@ enum EncodeErrorKind {
 /// pointer.
 #[derive(Debug)]
 pub struct BadInstruction {
-    ip: usize,
+    pub(crate) ip: usize,
 }
 
 impl fmt::Display for BadInstruction {
@@ -256,7 +179,7 @@ impl error::Error for BadInstruction {}
 /// pointer.
 #[derive(Debug)]
 pub struct BadJump {
-    jump: usize,
+    pub(crate) jump: usize,
 }
 
 impl fmt::Display for BadJump {
